services/user_service/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy[asyncio]
asyncmy
python-dotenv
python-multipart
redis
bcrypt
passlib[bcrypt]